+++ daemon-console/Models/OCR/ReadResult.cs
@@ -14,10 +14,10 @@ namespace daemon_console.Models.OCR
         public double Angle { get; set; }
 
         [JsonPropertyName("width")]
-        public int Width { get; set; }
+        public float Width { get; set; }
 
         [JsonPropertyName("height")]
-        public int Height { get; set; }
+        public float Height { get; set; }
 
         [JsonPropertyName("unit")]
         public string Unit { get; set; }
